Move functions to random_utils
